SelectPatternView afgemaakt
misschien moet er nog wat kleins met kleurtjes gebeuren... maar voor de rest is dit compleet.
hier staat ook een beginnetje in voor de dieView. namelijk PaneWithEyes.
PaneWithEyes is een pane met de ogen van een dobbelsteen. die moet je dan overerven etc.
